refactor: add package,  change packaging
- add constants
- add domain
- 이름에 맞는 패키지로 클래스이동
